chore: folder sturcture & move zod validation to server action
