+++ app/(auth)/login/page.tsx
@@ -1,5 +1,5 @@
-import { LoginForm } from '@/app/(auth)/components';
-import { login } from '@/app/(auth)/actions';
+import { LoginForm } from '@/app/(auth)/_components/LoginForm';
+import { login } from '@/app/(auth)/_actions/login';
 
 export default function LoginPage() {
   return <LoginForm onSubmitAction={login} />;
