+++ app/(auth)/register/page.tsx
@@ -1,5 +1,5 @@
-import { RegisterForm } from '@/app/(auth)/components';
-import { register } from '@/app/(auth)/actions';
+import { RegisterForm } from '@/app/(auth)/_components/RegisterForm';
+import { register } from '@/app/(auth)/_actions/register';
 
 export default function RegisterPage() {
   return <RegisterForm onSubmitAction={register} />;
